Manually add URLQueryItems
